[mod] Using generic class_based views
